Optimization
see https://gist.github.com/eteubert/4054e4953104c9f84e90

+++ lib/modules/social/model/show_service.php
@@ -18,4 +18,4 @@ ShowService::property( 'id', 'INT NOT NULL AUTO_INCREMENT PRIMARY KEY' );
 ShowService::property( 'service_id', 'INT' );
 ShowService::property( 'value', 'TEXT' );
 ShowService::property( 'title', 'TEXT' );
-ShowService::property( 'position', 'INT' );
+ShowService::property( 'position', 'FLOAT' );
